requirements.txt
flask==1.0.2
numpy==1.16.1
opencv-python==4.0.0.21
pyzmq==17.1.2
